+++ src/Syntax/Resolve.hs
@@ -20,11 +20,14 @@ import Control.Applicative
 import Data.Traversable
 import Data.Semigroup
 import Data.Foldable
+import Data.Spanned
 
 import Syntax.Resolve.Scope
 import Syntax.Resolve.Toplevel
+import Syntax.Pretty
 import Syntax.Subst
-import Syntax
+
+import Pretty
 
 data ResolveError
   = NotInScope (Var Parsed)
@@ -36,6 +39,15 @@ data ResolveError
   | ArisingFrom ResolveError (Expr Parsed)
   | ArisingFromTop ResolveError (Toplevel Parsed)
   deriving (Eq, Ord, Show)
+
+instance Pretty ResolveError where
+  pretty (NotInScope e) = string "Variable not in scope:" <> verbatim e
+  pretty (NoSuchModule e) = string "Module not in scope:" <> verbatim e
+  pretty (Ambiguous v _) = string "Ambiguous reference to variable:" <> verbatim v
+  pretty (EmptyMatch e) = pretty (annotation e) <> string ": Empty match expression"
+  pretty (EmptyBegin e) = pretty (annotation e) <> string ": Empty begin expression"
+  pretty (ArisingFrom er ex) = pretty (annotation ex) <> colon <+> pretty er <#> indent 2 (bullet (string "Arising from use of ") <+> verbatim ex)
+  pretty (ArisingFromTop er ex) = pretty (annotation ex) <> colon <+> pretty er <#> indent 2 (bullet (string "Arising in") <+> verbatim ex)
 
 type MonadResolve m = (MonadError ResolveError m, MonadReader Scope m, MonadGen Int m, MonadState ModuleScope m)
